# spec/openapi.yaml
openapi: 3.1.0
info:
  title: arw-server
  description: Unified ARW server API surface (headless-first).
  license:
    name: MIT OR Apache-2.0
    identifier: MIT OR Apache-2.0
  version: 0.1.4
paths:
...
  /admin/feedback/analyze:
    post:
      tags:
      - Admin/Feedback
      operationId: feedback_analyze_doc
      responses:
        '200':
          description: Recomputed suggestions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackState'
        '401':
          description: Unauthorized
      description: 'Feedback: POST /admin/feedback/analyze.'
  /admin/feedback/apply:
    post:
      tags:
      - Admin/Feedback
      operationId: feedback_apply_doc
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackApplyRequest'
        required: true
      responses:
        '200':
          description: Suggestion applied
          content:
            application/json:
              schema: {}
        '400':
          description: Invalid suggestion payload
        '401':
          description: Unauthorized
        '403':
          description: Policy denied
        '404':
          description: Unknown suggestion
      description: 'Feedback: POST /admin/feedback/apply.'
  /admin/feedback/auto:
    post:
      tags:
      - Admin/Feedback
      operationId: feedback_auto_doc
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackAutoRequest'
        required: true
      responses:
        '200':
          description: Auto-apply updated
          content:
            application/json:
              schema: {}
        '401':
          description: Unauthorized
      description: 'Feedback: POST /admin/feedback/auto.'
  /admin/feedback/policy:
    get:
      tags:
      - Admin/Feedback
      operationId: feedback_policy_doc
      responses:
        '200':
          description: Effective policy
          content:
            application/json:
              schema: {}
        '401':
          description: Unauthorized
      description: 'Feedback: GET /admin/feedback/policy.'
  /admin/feedback/reset:
    post:
      tags:
      - Admin/Feedback
      operationId: feedback_reset_doc
      responses:
        '200':
          description: Feedback state cleared
          content:
            application/json:
              schema: {}
        '401':
          description: Unauthorized
      description: 'Feedback: POST /admin/feedback/reset.'
  /admin/feedback/rollback:
    post:
      tags:
      - Admin/Feedback
      operationId: feedback_rollback_doc
      parameters:
      - name: to
        in: query
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      responses:
        '200':
          description: Snapshot restored
          content:
            application/json:
              schema: {}
        '401':
          description: Unauthorized
        '404':
          description: Snapshot not found
      description: 'Feedback: POST /admin/feedback/rollback.'
  /admin/feedback/signal:
    post:
      tags:
      - Admin/Feedback
      operationId: feedback_signal_doc
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackSignalRequest'
        required: true
      responses:
        '200':
          description: Signal recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackState'
        '401':
          description: Unauthorized
      description: 'Feedback: POST /admin/feedback/signal.'
  /admin/feedback/state:
    get:
      tags:
      - Admin/Feedback
      operationId: feedback_state_doc
      responses:
        '200':
          description: Feedback engine state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackState'
        '401':
          description: Unauthorized
      description: 'Feedback: GET /admin/feedback/state.'
  /admin/feedback/suggestions:
    get:
      tags:
      - Admin/Feedback
      operationId: feedback_suggestions_doc
      responses:
        '200':
          description: Current suggestions
          content:
            application/json:
              schema: {}
        '401':
          description: Unauthorized
      description: 'Feedback: GET /admin/feedback/suggestions.'
  /admin/feedback/updates:
    get:
      tags:
      - Admin/Feedback
      operationId: feedback_updates_doc
      parameters:
      - name: since
        in: query
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      responses:
        '200':
          description: Suggestions updated
          content:
            application/json:
              schema: {}
        '204':
          description: No changes since provided version
        '401':
          description: Unauthorized
      description: 'Feedback: GET /admin/feedback/updates.'
  /admin/feedback/versions:
    get:
      tags:
      - Admin/Feedback
      operationId: feedback_versions_doc
      responses:
        '200':
          description: Available snapshots
          content:
            application/json:
              schema: {}
        '401':
          description: Unauthorized
      description: 'Feedback: GET /admin/feedback/versions.'
  /admin/goldens/add:
    post:
      tags:
      - Admin/Experiments
      operationId: goldens_add_doc
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoldensAddRequest'
        required: true
      responses:
        '200':
          description: Golden added
          content:
            application/json:
              schema: {}
        '400':
          description: Persist failed
        '401':
          description: Unauthorized
      description: 'Experiments: POST /admin/goldens/add.'
  /admin/goldens/list:
    get:
      tags:
      - Admin/Experiments
      operationId: goldens_list_doc
      parameters:
      - name: proj
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Project goldens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoldensListResponse'
        '401':
          description: Unauthorized
      description: 'Experiments: GET /admin/goldens/list.'
  /admin/goldens/run:
    post:
      tags:
      - Admin/Experiments
      operationId: goldens_run_doc
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoldensRunRequest'
        required: true
      responses:
        '200':
          description: Evaluation summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvalSummary'
        '401':
          description: Unauthorized
      description: 'Experiments: POST /admin/goldens/run.'
...
components:
  schemas:
...
    EvalResultItem:
      type: object
      required:
      - id
      - ok
      - latency_ms
      properties:
        id:
          type: string
        latency_ms:
          type: integer
          format: int64
          minimum: 0
        note:
          type:
          - string
          - 'null'
        ok:
          type: boolean
    EvalSummary:
      type: object
      required:
      - total
      - passed
      - failed
      - avg_latency_ms
      - items
      properties:
        avg_ctx_items:
          type: integer
          format: int64
          minimum: 0
        avg_ctx_tokens:
          type: integer
          format: int64
          minimum: 0
        avg_latency_ms:
          type: integer
          format: int64
          minimum: 0
        failed:
          type: integer
          minimum: 0
        items:
          type: array
          items:
            $ref: '#/components/schemas/EvalResultItem'
        passed:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
...
    FeedbackApplyRequest:
      type: object
      required:
      - id
      properties:
        id:
          type: string
    FeedbackAutoRequest:
      type: object
      required:
      - enabled
      properties:
        enabled:
          type: boolean
    FeedbackSignal:
      type: object
      required:
      - id
      - ts
      - kind
      - target
      - confidence
      - severity
      properties:
        confidence:
          type: number
          format: double
        id:
          type: string
        kind:
          type: string
        note:
          type:
          - string
          - 'null'
        severity:
          type: integer
          format: int32
          minimum: 0
        target:
          type: string
        ts:
          type: string
    FeedbackSignalRequest:
      type: object
      required:
      - kind
      - target
      - confidence
      - severity
      properties:
        confidence:
          type: number
          format: double
        kind:
          type: string
        note:
          type:
          - string
          - 'null'
        severity:
          type: integer
          format: int32
          minimum: 0
        target:
          type: string
    FeedbackState:
      type: object
      properties:
        auto_apply:
          type: boolean
        signals:
          type: array
          items:
            $ref: '#/components/schemas/FeedbackSignal'
        suggestions:
          type: array
          items:
            $ref: '#/components/schemas/Suggestion'
    GoldenItem:
      type: object
      required:
      - id
      - kind
      properties:
        expect: {}
        id:
          type: string
        input: {}
        kind:
          type: string
    GoldensAddRequest:
      type: object
      required:
      - proj
      - kind
      properties:
        expect: {}
        id:
          type:
          - string
          - 'null'
        input: {}
        kind:
          type: string
        proj:
          type: string
    GoldensListResponse:
      type: object
      required:
      - project
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GoldenItem'
        project:
          type: string
    GoldensRunRequest:
      type: object
      required:
      - proj
      properties:
        compression_aggr:
          type:
          - number
          - 'null'
          format: double
        limit:
          type:
          - integer
          - 'null'
          minimum: 0
        mmr_lambda:
          type:
          - number
          - 'null'
          format: double
        proj:
          type: string
        retrieval_k:
          type:
          - integer
          - 'null'
          minimum: 0
        temperature:
          type:
          - number
          - 'null'
          format: double
        vote_k:
          type:
          - integer
          - 'null'
          minimum: 0
...
    Suggestion:
      type: object
      required:
      - id
      - action
      - params
      properties:
        action:
          type: string
        confidence:
          type: number
          format: double
        id:
          type: string
        params: {}
        rationale:
          type: string
...